move _env to lib/env
