+++ sites/svelte.dev/src/routes/blog/index.svelte
@@ -1,5 +1,5 @@
 <script context="module">
-	import { API_BASE } from '../../_env';
+	import { API_BASE } from '$lib/env';
 
 	export async function load({ fetch }) {
 		const res = await fetch(`${API_BASE}/docs/svelte/blog`);
