+++ sites/svelte.dev/src/routes/examples/__layout.svelte
@@ -1,14 +1,14 @@
 <script context="module">
-	import {API_BASE} from '../../_env';
+	import { API_BASE } from '$lib/env';
 
 	export async function load() {
 		const examples = await fetch(`${API_BASE}/docs/svelte/examples`).then(r => r.json());
 
-		return { 
+		return {
 			props: {
 				examples
 			},
-			maxage: 60 
+			maxage: 60
 		};
 	}
 </script>
