+++ sites/svelte.dev/src/routes/faq/index.svelte
@@ -1,9 +1,9 @@
 <script context="module">
-	import { API_BASE } from '../../_env';
+	import { API_BASE } from '$lib/env';
 
 	export async function load({ fetch }) {
 		const faqs = await fetch(`${API_BASE}/docs/svelte/faq?content`).then(r => r.json());
-		
+
 		return {
 			props: { faqs },
 			maxage: 60
