+++ sites/svelte.dev/src/utils/highlight.js
@@ -6,11 +6,9 @@ import 'prism-svelte';
 
 export function highlight(source, lang) {
 	const plang = langs[lang] || '';
-	const highlighted = plang ? PrismJS.highlight(
-		source,
-		PrismJS.languages[plang],
-		lang,
-	) : source.replace(/[&<>]/g, c => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;' })[c]);
+	const highlighted = plang
+		? PrismJS.highlight(source, PrismJS.languages[plang], lang)
+		: source.replace(/[&<>]/g, c => ({ '&': '&amp;', '<': '&lt;', '>': '&gt;' }[c]));
 
 	return `<pre class='language-${plang}'><code>${highlighted}</code></pre>`;
 }
